apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Authentication dependencies
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2


# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
passlib[bcrypt]==1.7.4

